+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.0.0
 
 - update @vcmap/core and @vcmap/ui to 6.x
+- fix windowTitle update, feature actions update and German category name in MyWorkspace